added override of default bytesPerChunk for large files
This won't change the behaviour when using inline args, but it will let
the program automatically set the chunk size to about 1/20 the size of
the file if the file is larger than 2000 bytes. This will prevent the
slow down that I explained in slack when using default settings for
bytesPerChunk in sender.