Work on Arguments (for sender but we can reuse a lot for receiver)
I created an abstract Arg class and extends it in a bunch of the inline
arguments we will use. Each argument has a flag string like "-f" that
indicate the following arg inline is for that argument. I have a special
-help flag that will trigger a help output for the program rather than
running the program. You can access the argument's stored value by using
the getValue() method. each argument has a default value that it will
use if there is nothing entered in the main's args String array.