Work on the receiver
I added a way for us to have the sender pass its arguments to the
receiver and proxy, but we need to set up the proxy to use it.

I did a bit of work on the receiver and ended up changing a little bit
in how the sender does things.

I added a few methods to the Frame classes including a getLength()
instance method and a getSequenceNumber() method in the ChunkFrame
class.